requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1
factory-boy==3.2.0
Pillow~=8.3.1